convert recommendation.id to Long, convert primary product to String

+++ features/discovery/home_recom/src/main/java/com/tokopedia/home_recom/model/datamodel/RecommendationCarouselDataModel.kt
@@ -26,5 +26,5 @@ class RecommendationCarouselDataModel(
 
     fun contains(item: RecommendationItem) = products.any { it.productItem.productId == item.productId }
 
-    fun contains(id: Int) = products.any { it.productItem.productId == id }
+    fun contains(id: Long) = products.any { it.productItem.productId == id }
 }
