+++ features/discovery/home_recom/src/main/java/com/tokopedia/home_recom/util/RecommendationMapperExt.kt
@@ -45,7 +45,7 @@ fun List<RecommendationWidget>.mapDataModel(): List<HomeRecommendationDataModel>
 
 fun ProductInfoDataModel.mapToRecommendationTracking(): RecommendationItem{
     return RecommendationItem(
-            productId = productDetailData?.id ?: -1,
+            productId = (productDetailData?.id ?: -1).toLong(),
             position = 0,
             name = productDetailData?.name ?: "",
             appUrl = productDetailData?.appUrl ?: "",
