+++ features/discovery/home_wishlist/src/main/java/com/tokopedia/home_wishlist/view/fragment/WishlistFragment.kt
@@ -158,7 +158,7 @@ open class WishlistFragment : Fragment(), WishlistListener, TopAdsListener {
                 val wishlistStatusFromPdp = data.getBooleanExtra(WIHSLIST_STATUS_IS_WISHLIST,
                         false)
                 viewModel.onPDPActivityResultForWishlist(
-                        id.toInt(),
+                        id.toLong(),
                         wishlistStatusFromPdp
                 )
             }
