+++ features/discovery/recommendation_widget_common/src/main/java/com/tokopedia/recommendation_widget_common/presentation/model/RecommendationItem.kt
@@ -3,7 +3,7 @@ package com.tokopedia.recommendation_widget_common.presentation.model
 import com.tokopedia.kotlin.model.ImpressHolder
 
 data class RecommendationItem(
-        val productId: Int = 0,
+        val productId: Long = 0,
         val name: String = "",
         val categoryBreadcrumbs: String = "",
         val url: String = "",
